==> 05_Feature Engineering.ipynb <==
# Feature Enginnering
- "Peak Signal Characterisation" Spalten werden auf "Modality","Tail" und "Shoulder" reduziert
- Polymer wird aus der "Name of Original DSC Measurement" Spalte etrahiert

import pandas as pd
import re

file_path_data = "data/03_Data_2nd_Heating.csv"
df_data = pd.read_csv(file_path_data)
df_data

# Extracting distinct values from columns related to "Peak Signal Characterisation"
peak_signal_columns = [col for col in df_data.columns if "Peak Signal Characterisation" in col]
distinct_values = {col: df_data[col].unique().tolist() for col in peak_signal_columns}

# Output the distinct values
for column, values in distinct_values.items():
    print(f"{column}: {values}")

def sort_columns(df, last_columnname='Aging Level'):
    """
    Sortiert die Spalten eines DataFrames in folgender Reihenfolge:
    1. Spalten, die nicht mit 'T' oder 'Peak Signal' beginnen und nicht die letzte Spalte sind.
    2. Spalten, die mit 'T' beginnen.
    3. Spalten, die mit 'Peak Signal' beginnen.
    4. Die letzte Spalte (standardmäßig 'Aging Level').

    Parameters:
        df (pd.DataFrame): Der zu sortierende DataFrame.
        last_columnname (str): Name der letzten Spalte (Standard: 'Aging Level').

    Returns:
        pd.DataFrame: DataFrame mit neu sortierten Spalten.
    """
    columns = df.columns
    characterisation=["Modality", "Tail", "Shoulder"]

    # Listen für die Sortierung
    first_columns = [col for col in columns if not col.startswith(('T','H', 'Peak Signal')) and col != last_columnname and col not in characterisation]
    t_columns = [col for col in columns if col.startswith('T') and col not in characterisation]
    h_columns = [col for col in columns if col.startswith('H')]
    peak_columns = [col for col in columns if col.startswith('Peak Signal')]
    last_column = [last_columnname] if last_columnname in columns else []

    # Neu sortieren: erst 'first_columns', dann 't_columns', dann 'peak_columns', dann 'last_column'
    sorted_columns = first_columns + t_columns + h_columns + peak_columns + characterisation +  last_column

    # DataFrame mit neu sortierten Spalten zurückgeben
    return df[sorted_columns]

### a) Feature Consolidation  - Peak Signal Characterisation

# Funktion zur Festlegung der Peak Signal Characterisation
def identify_characterisation_row(row):
    
    # Helper-Funktion, um zu prüfen, ob ein Wert als "nicht 0" gilt
    def is_not_zero(value):
        return str(value) != '0' and value != 0 and value != 0.0

    # Modality bestimmen
    modality_columns = {
        'bimodal': is_not_zero(row['Peak Signal Characterisation_Bimodal']),
        'sharp': is_not_zero(row['Peak Signal Characterisation_Sharp Monodal']),
        'wide': is_not_zero(row['Peak Signal Characterisation_Wide Monodal']),
        'quasi': is_not_zero(row['Peak Signal Characterisation_Quasibimodal'])
    }
    
     # Wenn kein Wert einen Eintrag hat oder mehrere Werte vorliegen, dann 'undefined'
    if sum(modality_columns.values()) == 1:
        modality = [key for key, value in modality_columns.items() if value][0]
    else:
        modality = 'undefined'

    # Tail bestimmen
    if is_not_zero(row['Peak Signal Characterisation_Left Tail']):
        tail = 'left'
    elif is_not_zero(row['Peak Signal Characterisation_Right Tail']):
        tail = 'right'
    else:
        tail = 'undefined'

    # Shoulder bestimmen
    if is_not_zero(row['Peak Signal Characterisation_Left Shoulder']):
        shoulder = 'left'
    elif is_not_zero(row['Peak Signal Characterisation_Right Shoulder']):
        shoulder = 'right'
    else:
        shoulder = 'undefined'

    return pd.Series([modality, tail, shoulder], index=['Modality', 'Tail', 'Shoulder'])



# Funktion zur Transformation des gesamten DataFrames
def transform_peak_signal_characteristics(df):
    # Transformation anwenden
    df[['Modality', 'Tail', 'Shoulder']] = df.apply(identify_characterisation_row, axis=1)
    
    # Alle Spalten mit "Peak Signal Characterisation" löschen
    columns_to_drop = [col for col in df.columns if 'Peak Signal Characterisation' in col]
    df.drop(columns=columns_to_drop, inplace=True)
    
    return df

# Transformation anwenden
df_transformed = transform_peak_signal_characteristics(df_data)
df_transformed

# Häufigkeiten der Werte für jede Spalte berechnen und einzeln anzeigen
print(f"{df_transformed['Modality'].value_counts()}\n")
print(f"{df_transformed['Tail'].value_counts()}\n")
print(f"{df_transformed['Shoulder'].value_counts()}\n")

### b) Feature Generation - Hm (y-Wert zu der Temperatur Tm) 

df_experiments = pd.read_csv("data/Experiments_2nd_Heating.csv", index_col = 0)
df_experiments

# Spalte "Tm in °C" in Integer (gerundet) konvertieren
df_transformed["Tm"] = df_transformed["Tm in °C"].round().astype(int)
df_transformed

count = 0

# Leere Spalte "Hm" initialisieren
df_transformed["Hm"] = None

# Iterative Schleife, um Werte zuzuweisen
for i, row in df_transformed.iterrows():
    temp = row["Tm"]  # Temperatur
    measurement_name = row["Name of Original DSC Measurement"]  # Name der Messung
    
    # Überprüfung der Index- und Spaltenexistenz
    temp_exists = temp in df_experiments.index
    column_exists = measurement_name in df_experiments.columns

    if temp_exists and column_exists:
        # Wert zuweisen
        df_transformed.at[i, "Hm"] = df_experiments.at[temp, measurement_name]
    else:
        print(f"{measurement_name} fehlt in den Messwerten")
        count+=1

print(f"\nEs fehlen ingesamt {count} Messwerte")

# Entfernen der Zeilen mit None/NaN-Werten in der Spalte "Hm"
df_transformed = df_transformed.dropna(subset=["Hm"])
df_transformed

### c) Feature Extraction - Polymer

df_transformed["Name of Original DSC Measurement"]

df_transformed['Polymer'] = df_transformed['Name of Original DSC Measurement'].apply(
    lambda x: re.search(r'-(.*?)-', x).group(1) if re.search(r'-(.*?)-', x) else None
)

df_transformed['Polymer'].value_counts()

df_transformed = sort_columns(df_transformed)
df_transformed

### Export

df_transformed.to_csv("data/050_Data_2st_Heating.csv", index=False)

## Feature-based Splitting nach Polymer

### PET (Polyethylenterephthalat)

# Filtern der Zeilen, in denen Polymer gleich "PET" ist
pet_data = df_transformed[df_transformed['Polymer'] == 'PET']
pet_data

# Exportieren der gefilterten Daten als CSV
pet_data.to_csv("data/051_Data_2st_Heating_PET.csv", index=False)

### PE (Polyethylen)

# Filtern der Zeilen, in denen Polymer gleich "PE" ist
pe_data = df_transformed[df_transformed['Polymer'] == 'PE']
pe_data

# Exportieren der gefilterten Daten als CSV
pe_data.to_csv("data/052_Data_2st_Heating_PE.csv", index=False)
